refactor: divide into small components
